Reshape OCR result model: reading-order elements, doc-level coordinates, usage tokens
Continue the IOcrClient reshape ahead of API review (#7587/#7588).

- Replace parallel OcrPage.Blocks/Tables/Images with a single reading-order
  Elements list over a new polymorphic OcrElement base (OcrBlock/OcrTable/OcrImage
  derive); consumers project with OfType<T>().
- Add optional nested OcrTableCell.Elements for structured cell content; keep the
  Content string as a flat-text convenience.
- Convert OcrCoordinateUnit to a closed enum and add Point; add an
  OcrCoordinateOrigin enum; move both to document level (OcrResult/OcrPageResult)
  and remove the per-page CoordinateUnit.
- Keep OcrBlockKind and OcrTableCellKind as open structs; add RowHeader and
  RowSection well-known cell kinds.
- Add nullable Input/Output/TotalTokenCount to OcrUsage for the vision-LLM path.
- Rename Markdown to Text on OcrPage/OcrResult; drop response ModelId,
  OcrPage.Confidence, and OcrOptions.IncludeImages; FromRectangle takes float.
- Keep ToOcrResult/ToOcrResultAsync reducers and the convenience overloads as
  extension methods, matching the MEAI ChatResponseExtensions precedent.
- Regenerate the ApiChief baselines for the reshaped surface.

Copilot-Session: 19ea13dd-f444-4942-b751-162711478580

+++ src/Libraries/Microsoft.Extensions.AI.Abstractions/Ocr/OcrBoundingRegion.cs
@@ -45,13 +45,13 @@ public class OcrBoundingRegion
     /// <param name="right">The right coordinate.</param>
     /// <param name="bottom">The bottom coordinate.</param>
     /// <returns>A region whose polygon is the four corners of the rectangle.</returns>
-    public static OcrBoundingRegion FromRectangle(int pageNumber, double left, double top, double right, double bottom)
+    public static OcrBoundingRegion FromRectangle(int pageNumber, float left, float top, float right, float bottom)
         => new(pageNumber,
         [
-            new OcrPoint((float)left, (float)top),
-            new OcrPoint((float)right, (float)top),
-            new OcrPoint((float)right, (float)bottom),
-            new OcrPoint((float)left, (float)bottom),
+            new OcrPoint(left, top),
+            new OcrPoint(right, top),
+            new OcrPoint(right, bottom),
+            new OcrPoint(left, bottom),
         ]);
 
     /// <summary>Computes the axis-aligned bounds of the polygon.</summary>
